move notification code into its own file under src/ui
